+++ conf/docker/entrypoint.sh
@@ -8,13 +8,7 @@ export DJANGO_SETTINGS_MODULE=run.settings_docker
 
 python manage.py migrate --run-syncdb --fake-initial
 
-python manage.py shell <<EOL
-from glifestream.stream.models import Service
-if not Service.objects.exists():
-    print('Loading initial data...')
-    from django.core.management import call_command
-    call_command('loaddata', 'glifestream/stream/fixtures/initial_data.json')
-EOL
+python manage.py load_initial_data
 
 python manage.py collectstatic --no-input
 python worker.py --init-files-dirs